requirements.txt
annotated-types==0.7.0
et_xmlfile==2.0.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scipy==1.15.3
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
